refactor: move results-related components in a subfolder + comment console log
